+++ src/utils/mod.rs
@@ -1,3 +1,5 @@
+pub mod filters;
 pub mod path_utils;
 
+pub use filters::match_filter;
 pub use path_utils::expand_path_str;